demo/requirements.txt
azure-cognitiveservices-vision-customvision
azure-common
azure-core
faiss-cpu
Flask
huggingface-hub
langchain
nltk
numpy
openai
openapi-schema-pydantic
opencv-python
pandas
pypdf
Pillow
scikit-learn
scipy
tiktoken

